+++ inc/strategy/StrategyImpl.h
@@ -7,7 +7,7 @@ class StrategyImpl
 {
 public:
     virtual Move checkOrBet() = 0;
-    virtual Move callRaiseOrFold(unsigned int bb) = 0;
+    virtual Move callRaiseOrFold(float bb) = 0;
     virtual ~StrategyImpl() = default;
 };
